- failingStreak type changed from int to long - introduced TestcontainersHealthStates enum as a wrapper to hide Docker.DotNet.Models type - added two new properties to expose health state and health check failing streak

+++ src/Testcontainers/Configurations/WaitStrategies/IWaitForContainerOS.cs
@@ -82,7 +82,7 @@ namespace DotNet.Testcontainers.Configurations
     /// <returns>A configured instance of <see cref="IWaitForContainerOS" />.</returns>
     /// <param name="failingStreak">Number of tolerated failed attempts before throwing a <see cref="ContainerDidNotStartException"/>, default value is 20.</param>
     [PublicAPI]
-    IWaitForContainerOS UntilContainerIsHealthy(int failingStreak = 20);
+    IWaitForContainerOS UntilContainerIsHealthy(long failingStreak = 20);
 
     /// <summary>
     /// Returns a collection with all configured wait strategies.
